Refactor and Fix Various Issues in Philosophers Simulation
Refactored Code:

Added clean_data to ensure proper cleanup on errors.
Introduced new think routine for clarity and modularity.
Improved logic in dinner_start and dinner_routine.
Fixes:

Corrected pointer handling for last_meal_time in set_long.
Replaced incorrect pthread_mutex_destory with pthread_mutex_destroy.
Fixed type mismatches (e.g., int to size_t) for safer operations.
Enhancements:

Added colorful status logs (Started parsing, Completed initialization, etc.) for better debugging and simulation flow tracking.
Updated printing functions to include additional simulation information.
Updated Headers:

Adjusted meals_eaten and last_meal_time types in s_philo.
Included sys/time.h in structs.h to support time-related operations.

+++ philo/incl/structs.h
@@ -6,7 +6,7 @@
 /*   By: lade-kon <lade-kon@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/10/11 16:22:56 by lade-kon      #+#    #+#                 */
-/*   Updated: 2024/11/28 16:16:40 by lade-kon      ########   odam.nl         */
+/*   Updated: 2024/12/06 14:55:20 by lade-kon      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,10 @@
 # include <pthread.h>
 # include <stdbool.h>
 # include <inttypes.h>
+# include <sys/time.h>
 
 typedef struct s_table	t_table;
+
 
 /**
  * @brief PHILO
@@ -29,8 +31,8 @@ typedef struct s_table	t_table;
 typedef struct s_philo
 {
 	int				philo_id;
-	int				meals_eaten;
-	size_t			last_meal_time; //time passed from last meal
+	size_t			meals_eaten;
+	long			last_meal_time; //time passed from last meal
 	bool			full;
 	pthread_mutex_t	*first_fork;
 	pthread_mutex_t	*second_fork;
